Adjust packaging to consolidate types
