+++ platforms/software/dependency-management/src/main/java/org/gradle/internal/component/resolution/failure/formatting/CapabilitiesDescriber.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.internal.component.resolution.failure;
+package org.gradle.internal.component.resolution.failure.formatting;
 
 import org.gradle.api.capabilities.Capability;
